Lots of changes!
This commit was made to add the pictures to the boxes in the board after the ship placement phase. There were changes to boat_selection because I needed to update the player arrays and call functions to begin the combat_phase. Lastly, I added in the photos that are needed to display hits, misses, and the covers for unselected boxes